// src/types/database.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      vegetables: {
        Row: {
          category: string;
          created_at: string;
          description: string;
          id: string;
          image_path: string | null;
          name: string;
          prep_guide: string | null;
          season: string | null;
          storage_guide: string | null;
        };
        Insert: {
          category: string;
          created_at?: string;
          description: string;
          id: string;
          image_path?: string | null;
          name: string;
          prep_guide?: string | null;
          season?: string | null;
          storage_guide?: string | null;
        };
        Update: {
          category?: string;
          created_at?: string;
          description?: string;
          id?: string;
          image_path?: string | null;
          name?: string;
          prep_guide?: string | null;
          season?: string | null;
          storage_guide?: string | null;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};